@@ -18,7 +18,7 @@
 // https://github.com/fabianfett/pure-swift-json/blob/master/Sources/PureSwiftJSONParsing/DocumentReader.swift
 
 /// Reader object for parsing String buffers
-public struct HBParser {
+public struct HBParser: Sendable {
     public enum Error: Swift.Error {
         case overflow
         case unexpected
@@ -668,7 +668,3 @@ extension Set where Element == Unicode.Scalar {
         self = Set(string.unicodeScalars)
     }
 }
-
-#if swift(>=5.6)
-extension HBParser: Sendable {}
-#endif
